+++ packages/lumx-react/src/components/dialog/Dialog.tsx
@@ -1,6 +1,6 @@
-import React, { Children, ReactElement, ReactNode, Ref, RefObject, useMemo, useRef, useState } from 'react';
+import React, { Children, ReactElement, Ref, RefObject, useMemo, useRef, useState } from 'react';
 
-import { HeadingLevelProvider, Progress, ProgressVariant, Size } from '@lumx/react';
+import { HeadingLevelProvider, ProgressCircular, Size } from '@lumx/react';
 
 import { DIALOG_TRANSITION_DURATION, DOCUMENT } from '@lumx/react/constants';
 import { useCallbackOnEscape } from '@lumx/react/hooks/useCallbackOnEscape';
@@ -9,34 +9,27 @@ import { useIntersectionObserver } from '@lumx/react/hooks/useIntersectionObserv
 
 import { GenericProps, HasCloseMode, isComponent } from '@lumx/react/utils/type';
 import { partitionMulti } from '@lumx/react/utils/partitionMulti';
-import type { LumxClassName } from '@lumx/core/js/types';
-import { classNames } from '@lumx/core/js/utils';
 import { ClickAwayProvider } from '@lumx/react/utils/ClickAwayProvider';
-import { mergeRefs } from '@lumx/react/utils/react/mergeRefs';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
+import { mergeRefs } from '@lumx/react/utils/react/mergeRefs';
 
 import { useDisableBodyScroll } from '@lumx/react/hooks/useDisableBodyScroll';
 import { useTransitionVisibility } from '@lumx/react/hooks/useTransitionVisibility';
 import { ThemeProvider } from '@lumx/react/utils/theme/ThemeContext';
 
 import { Portal } from '@lumx/react/utils';
+import {
+    Dialog as UI,
+    CLASSNAME,
+    COMPONENT_NAME,
+    DialogSizes,
+    BaseDialogProps as UIProps,
+} from '@lumx/core/js/components/Dialog';
 
 /**
  * Defines the props of the component.
  */
-export interface DialogProps extends GenericProps, HasCloseMode {
-    /** Footer content. */
-    footer?: ReactNode;
-    /** Whether the divider between the dialog content and the footer is always displayed (instead of showing it on scroll). */
-    forceFooterDivider?: boolean;
-    /** Header content. */
-    header?: ReactNode;
-    /** Whether the divider between the dialog content and the footer is always displayed (instead of showing it on scroll). */
-    forceHeaderDivider?: boolean;
-    /** Whether the indefinite progress indicator over the dialog content is displayed or not. */
-    isLoading?: boolean;
-    /** Whether the component is open or not. */
-    isOpen?: boolean;
+export interface DialogProps extends GenericProps, HasCloseMode, UIProps {
     /** Reference to the parent element that triggered modal opening (will get back focus on close). */
     parentElement?: RefObject<HTMLElement>;
     /** Reference to the dialog content element. */
@@ -65,21 +58,8 @@ export interface DialogProps extends GenericProps, HasCloseMode {
     children?: React.ReactNode;
 }
 
-export type DialogSizes = Extract<Size, 'tiny' | 'regular' | 'big' | 'huge'>;
-
 const isHeader = isComponent('header');
 const isFooter = isComponent('footer');
-
-/**
- * Component display name.
- */
-const COMPONENT_NAME = 'Dialog';
-
-/**
- * Component default class name and class prefix.
- */
-const CLASSNAME: LumxClassName<typeof COMPONENT_NAME> = 'lumx-dialog';
-const { block, element } = classNames.bem(CLASSNAME);
 
 /**
  * Component default props.
@@ -89,6 +69,8 @@ const DEFAULT_PROPS: Partial<DialogProps> = {
     size: Size.big,
     disableBodyScroll: true,
 };
+
+export { type DialogSizes };
 
 /**
  * Dialog component.
@@ -198,87 +180,43 @@ export const Dialog = forwardRef<DialogProps, HTMLDivElement>((props, ref) => {
 
     const isMounted = isOpen || isVisible || closeMode === 'hide';
 
-    return isMounted ? (
-        <Portal>
-            <div
-                ref={mergeRefs(rootRef, ref)}
-                {...forwardedProps}
-                className={classNames.join(
-                    className,
-                    block({
-                        'is-hidden': !isOpen,
-                        'is-loading': isLoading,
-                        'is-shown': isOpen || isVisible,
-                        [`size-${size}`]: Boolean(size),
-                    }),
-                )}
-                style={{ zIndex }}
-            >
-                <div className={element('overlay')} />
-
-                <HeadingLevelProvider level={2}>
-                    <ThemeProvider value={undefined}>
-                        <div className={element('container')} role="dialog" aria-modal="true" {...dialogProps}>
-                            <ClickAwayProvider
-                                callback={!shouldPreventCloseOnClickAway && onClose}
-                                childrenRefs={clickAwayRefs}
-                                parentRef={rootRef}
-                            >
-                                <section className={element('wrapper')} ref={wrapperRef}>
-                                    {(header || headerChildContent) && (
-                                        <header
-                                            {...headerChildProps}
-                                            className={classNames.join(
-                                                element('header', {
-                                                    'has-divider': Boolean(forceHeaderDivider || hasTopIntersection),
-                                                }),
-                                                headerChildProps?.className,
-                                            )}
-                                        >
-                                            {header}
-                                            {headerChildContent}
-                                        </header>
-                                    )}
-
-                                    <div ref={mergeRefs(contentRef, localContentRef)} className={element('content')}>
-                                        <div className={element('sentinel', { top: true })} ref={setSentinelTop} />
-
-                                        {content}
-
-                                        <div
-                                            className={element('sentinel', { bottom: true })}
-                                            ref={setSentinelBottom}
-                                        />
-                                    </div>
-
-                                    {(footer || footerChildContent) && (
-                                        <footer
-                                            {...footerChildProps}
-                                            className={classNames.join(
-                                                element('footer', {
-                                                    'has-divider': Boolean(forceFooterDivider || hasBottomIntersection),
-                                                }),
-                                                footerChildProps?.className,
-                                            )}
-                                        >
-                                            {footer}
-                                            {footerChildContent}
-                                        </footer>
-                                    )}
-
-                                    {isLoading && (
-                                        <div className={element('progress-overlay')}>
-                                            <Progress variant={ProgressVariant.circular} />
-                                        </div>
-                                    )}
-                                </section>
-                            </ClickAwayProvider>
-                        </div>
-                    </ThemeProvider>
-                </HeadingLevelProvider>
-            </div>
-        </Portal>
-    ) : null;
+    return isMounted
+        ? UI({
+              ClickAwayProvider,
+              HeadingLevelProvider,
+              Portal,
+              ThemeProvider,
+              className,
+              clickAwayRefs,
+              content,
+              contentRef: mergeRefs(localContentRef, contentRef),
+              dialogProps,
+              footer,
+              footerChildContent,
+              footerChildProps,
+              forceFooterDivider,
+              forceHeaderDivider,
+              handleClose: onClose,
+              hasBottomIntersection,
+              hasTopIntersection,
+              header,
+              headerChildContent,
+              ProgressCircular,
+              headerChildProps,
+              isLoading,
+              isOpen,
+              isVisible,
+              ref: mergeRefs(rootRef, ref),
+              rootRef,
+              setSentinelBottom,
+              setSentinelTop,
+              shouldPreventCloseOnClickAway,
+              size,
+              wrapperRef,
+              zIndex,
+              ...forwardedProps,
+          })
+        : null;
 });
 Dialog.displayName = COMPONENT_NAME;
 Dialog.className = CLASSNAME;