+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   `@lumx/core`:
+    -   Moved `Dialog` from `@lumx/react`
+
 ## [4.10.0][] - 2026-04-03
 
 ### Added